Add RecordData impl for A
