+++ crates/client/src/rr/mod.rs
@@ -27,12 +27,12 @@ pub use crate::proto::rr::resource;
 
 pub use self::dns_class::DNSClass;
 pub use self::record_data::RData;
-pub use self::record_data::RecordData;
 pub use self::record_type::RecordType;
 pub use self::resource::Record;
 pub use self::rr::domain::{IntoName, Label, Name};
 #[allow(deprecated)]
 pub use self::rr::IntoRecordSet;
+pub use self::rr::RecordData;
 pub use self::rr::RecordSet;
 
 /// All record data structures and related serialization methods
